Move gha into its own module
